ocr_worker/requirements.txt
packaging
pdf2image
pika
pillow
psycopg2-binary
pytesseract
python-dotenv
boto3
redis